Split hyperkube images (#173)
* Remove param kubernetesEndpoints

* Individual kube-proxy; extract kubelet/kubectl from kubeBinary

* Fix unit test

+++ vhdbuilder/packer/install-dependencies.sh
@@ -448,18 +448,29 @@ K8S_VERSIONS="
 1.18.6
 "
 for PATCHED_KUBERNETES_VERSION in ${K8S_VERSIONS}; do
-  HYPERKUBE_URL="mcr.microsoft.com/oss/kubernetes/hyperkube:v${PATCHED_KUBERNETES_VERSION}"
-  # NOTE: the KUBERNETES_VERSION will be used to tag the extracted kubelet/kubectl in /usr/local/bin
-  # it should match the KUBERNETES_VERSION format(just version number, e.g. 1.15.7, no prefix v)
-  # in installKubeletAndKubectl() executed by cse, otherwise cse will need to download the kubelet/kubectl again
-  KUBERNETES_VERSION=$(echo ${PATCHED_KUBERNETES_VERSION} | cut -d"_" -f1 | cut -d"-" -f1 | cut -d"." -f1,2,3)
-  # extractHyperkube will extract the kubelet/kubectl binary from the image: ${HYPERKUBE_URL}
-  # and put them to /usr/local/bin/kubelet-${KUBERNETES_VERSION}
-  extractHyperkube "docker"
-  # remove hyperkube here as the one that we really need is pulled later
-  docker image rm $HYPERKUBE_URL
+  if (($(echo ${PATCHED_KUBERNETES_VERSION} | cut -d"." -f2) < 17)); then
+    HYPERKUBE_URL="mcr.microsoft.com/oss/kubernetes/hyperkube:v${PATCHED_KUBERNETES_VERSION}"
+    # NOTE: the KUBERNETES_VERSION will be used to tag the extracted kubelet/kubectl in /usr/local/bin
+    # it should match the KUBERNETES_VERSION format(just version number, e.g. 1.15.7, no prefix v)
+    # in installKubeletAndKubectl() executed by cse, otherwise cse will need to download the kubelet/kubectl again
+    KUBERNETES_VERSION=$(echo ${PATCHED_KUBERNETES_VERSION} | cut -d"_" -f1 | cut -d"-" -f1 | cut -d"." -f1,2,3)
+    # extractHyperkube will extract the kubelet/kubectl binary from the image: ${HYPERKUBE_URL}
+    # and put them to /usr/local/bin/kubelet-${KUBERNETES_VERSION}
+    extractHyperkube "docker"
+    # remove hyperkube here as the one that we really need is pulled later
+    docker image rm $HYPERKUBE_URL
+  else
+    # strip the last .1 as that is for base image patch for hyperkube
+    if grep -iq hotfix <<< ${PATCHED_KUBERNETES_VERSION}; then
+      PATCHED_KUBERNETES_VERSION=`echo ${PATCHED_KUBERNETES_VERSION} | cut -d"." -f1,2,3,4`;
+    else
+      PATCHED_KUBERNETES_VERSION=`echo ${PATCHED_KUBERNETES_VERSION} | cut -d"." -f1,2,3`;
+    fi
+    KUBERNETES_VERSION=$(echo ${PATCHED_KUBERNETES_VERSION} | cut -d"_" -f1 | cut -d"-" -f1 | cut -d"." -f1,2,3)
+    extractKubeBinaries $KUBERNETES_VERSION "https://acs-mirror.azureedge.net/kubernetes/v${PATCHED_KUBERNETES_VERSION}/binaries/kubernetes-node-linux-amd64.tar.gz"
+  fi
 done
-ls -ltr /usr/local/bin >> ${VHD_LOGS_FILEPATH}
+ls -ltr /usr/local/bin/* >> ${VHD_LOGS_FILEPATH}
 
 # pull patched hyperkube image for AKS
 # this is used by kube-proxy and need to cover previously supported version for VMAS scale up scenario
@@ -489,7 +500,7 @@ PATCHED_HYPERKUBE_IMAGES="
 1.16.10-hotfix.20200714.1
 1.16.13-hotfix.20200714.1
 1.17.3-hotfix.20200601.1
-1.17.4.1
+1.17.4
 1.17.5.1
 1.17.7-hotfix.20200624
 1.17.7-hotfix.20200714.1
@@ -501,10 +512,27 @@ PATCHED_HYPERKUBE_IMAGES="
 1.18.6
 "
 for KUBERNETES_VERSION in ${PATCHED_HYPERKUBE_IMAGES}; do
-  CONTAINER_IMAGE="mcr.microsoft.com/oss/kubernetes/hyperkube:v${KUBERNETES_VERSION}"
-  pullContainerImage "docker" ${CONTAINER_IMAGE}
-  echo "  - ${CONTAINER_IMAGE}" >> ${VHD_LOGS_FILEPATH}
-  docker run --rm --entrypoint "" ${CONTAINER_IMAGE}  /bin/bash -c "iptables --version" | grep -v nf_tables && echo "Hyperkube contains no nf_tables" || (echo "Hyperkube contains nf_tables" && exit 99)
+  # TODO: after CCP chart is done, change below to get hyperkube only for versions less than 1.17 only
+  if (($(echo ${KUBERNETES_VERSION} | cut -d"." -f2) < 19)); then
+    CONTAINER_IMAGE="mcr.microsoft.com/oss/kubernetes/hyperkube:v${KUBERNETES_VERSION}"
+    pullContainerImage "docker" ${CONTAINER_IMAGE}
+    echo "  - ${CONTAINER_IMAGE}" >> ${VHD_LOGS_FILEPATH}
+    docker run --rm --entrypoint "" ${CONTAINER_IMAGE}  /bin/bash -c "iptables --version" | grep -v nf_tables && echo "Hyperkube contains no nf_tables" || (echo "Hyperkube contains nf_tables" && exit 99)
+  fi
+
+  # from 1.17 onwards start using kube-proxy as well
+  # strip the last .1 as that is for base image patch for hyperkube
+  if (($(echo ${KUBERNETES_VERSION} | cut -d"." -f2) >= 17)); then
+    if grep -iq hotfix <<< ${KUBERNETES_VERSION}; then
+      KUBERNETES_VERSION=`echo ${KUBERNETES_VERSION} | cut -d"." -f1,2,3,4`;
+    else
+      KUBERNETES_VERSION=`echo ${KUBERNETES_VERSION} | cut -d"." -f1,2,3`;
+    fi
+    CONTAINER_IMAGE="mcr.microsoft.com/oss/kubernetes/kube-proxy:v${KUBERNETES_VERSION}"
+    pullContainerImage "docker" ${CONTAINER_IMAGE}
+    docker run --rm --entrypoint "" ${CONTAINER_IMAGE}  /bin/bash -c "iptables --version" | grep -v nf_tables && echo "kube-proxy contains no nf_tables" || (echo "kube-proxy contains nf_tables" && exit 99)
+    echo "  - ${CONTAINER_IMAGE}" >>${VHD_LOGS_FILEPATH}
+  fi
 done
 
 ADDON_IMAGES="
